move template renderer to a separate file
