+++ packages/create-dara-app/create_dara_app/templates/default/{{ cookiecutter.__project_slug }}/{{ cookiecutter.__package_name }}/utils/template.py
@@ -1,0 +1,19 @@
+from dara.core.configuration import Configuration
+from dara.core.visual.components import Menu, RouterContent
+from dara.core.visual.components.sidebar_frame import SideBarFrame
+from dara.core.visual.template import TemplateBuilder
+
+def template_renderer(config: Configuration):
+    builder = TemplateBuilder(name='side-bar')
+
+    # Using the TemplateBuilder's helper method - add_router_from_pages
+    # to construct a router of page definitions
+    router = builder.add_router_from_pages(list(config.pages.values()))
+
+    builder.layout = SideBarFrame(
+        content=RouterContent(routes=router.content),
+        side_bar=Menu(routes=router.links),
+        logo_path='/static/dara_light.svg',
+    )
+
+    return builder.to_template()
